Change duration type from integer to floating point

+++ src/storage/tracks/schema.rs
@@ -56,7 +56,7 @@ table! {
         source_sync_rev_ordinal -> Nullable<BigInt>,
         source_sync_rev_timestamp -> Nullable<Timestamp>,
         content_type -> Text,
-        audio_duration_ms -> Nullable<BigInt>,
+        audio_duration_ms -> Nullable<Double>,
         audio_channels -> Nullable<SmallInt>,
         audio_samplerate_hz -> Nullable<Integer>,
         audio_bitrate_bps -> Nullable<Integer>,
